implemented separate obejct access rit for executing ad-hoc scripts (issue #3387); access rights attribute extended to 64 bit

+++ src/client/java/netxms-client/src/main/java/org/netxms/client/AccessListElement.java
@@ -34,7 +34,7 @@ public class AccessListElement extends AbstractAccessListElement
 	 * @param userId user id
 	 * @param accessRights bit mask
 	 */
-   public AccessListElement(int userId, int accessRights)
+   public AccessListElement(int userId, long accessRights)
 	{
 		super(userId, accessRights);
       inherited = false;
